pndarray restructure
